07_DeepLearning/Digit_Classification-acc-98.43/requirements.txt
torch
torchvision
streamlit
pillow
opencv-python-headless
streamlit-drawable-canvas
